+++ ts/src/sdto/index.ts
@@ -1,3 +1,3 @@
 export { FieldSpec, parseSchema } from "./FieldSpec";
 export { FluentAction, newAction, validateData } from "./FluentAction";
-export { SchemaBuilder, newSchemaBuilder, SupportedSignTypes } from "./SchemaBuilder";
+export { SchemaBuilder, newSchemaBuilder } from "./SchemaBuilder";